chore(ci): gate performance/scaling assertions behind a `perf` build tag
The release gate was going red for reasons unrelated to the diff:
TestLouvainScalingExponent fitted N^1.72 against a 1.45 bound on a loaded
Windows runner, TestRunSubprocess{Links,GroupAlgo}_CancelKillsGrandchildren
missed a 5s budget on macOS and passed on retry, and
TestIncremental_Performance_SingleFileEdit failed at 1.073s against a 1s
budget on an operation that takes 116-249ms in isolation. CI minutes are
free on a public repo, so the goal here is not speed — it is that a red
gate should mean something is broken.

Two separate mechanisms, because these are two separate problems.

1. Performance/scaling assertions move behind `//go:build perf`. These
   assert that something is FAST, or scales at exponent X, or allocates
   under N bytes — none of which a shared runner can measure. Moved:
   TestLouvainScalingExponent, TestBetweennessPerfBudget_28k,
   TestSampledBetweenness_ScratchIsReusedAcrossPivots,
   TestResponseShapePython_NearLinearScaling,
   TestBuildIndexFromModules_SubQuadratic, TestSingleRunPeak,
   TestOverlapPeak, TestDoctorQuickMode_Timing, TestSearchIndex_Scale_*,
   TestIncremental_Performance_SingleFileEdit. In every case the
   correctness half of the same change stays in the default suite; each
   perf file's header names which test carries it.

   The tag, not testing.Short(): it is explicit, cannot be enabled by
   accident, and is orthogonal to the 29 existing Short() guards — which
   must NOT be turned on in CI, see below.

2. Correctness-with-a-timeout tests stay in the gate, re-budgeted so they
   can only fail for their actual reason. These guard hangs, not latency:
   the pre-fix behaviour is 5s / 113s / forever, so the bound belongs
   between the two, not just above the healthy path.
     - cancel_processgroup_5999: 5s -> 30s unblock, 2s -> 10s reap
     - get_source_timeout: 100ms -> 1s, 500ms -> 2s (pre-#1773 = 5.000s)
     - find_zero_selectivity_5289: 2s -> 20s (unbounded scan = ~113s)
     - tier_watcher cold-wake: 500ms -> 5s (fake hook; guards blocking)
     - search_index concurrent: 2s -> 30s (kept for its -race value)

   Each was mutation-verified: reintroducing the original defect (or a
   stand-in for it) still fails the new bound.

Gated tests stay runnable and stay run: `make test-perf`, plus a weekly
+ dispatch `perf` workflow. pre-merge also gains `go vet -tags perf ./...`
so a tagged file that stops compiling is caught immediately rather than
up to a week later — gated tests that rot are worse than the flakes.

NOT doing: adding -short to CI. 29 files call testing.Short() and CI has
never passed -short, so no guard has ever fired. They cover the whole
internal/daemon/watch watcher suite, the internal/docgen LLM-mode
integration tests, internal/daemon/extract end-to-end equivalence, and
internal/extractors/golang large-file extraction — correctness, all of
it. -short would silently narrow the gate, which is the defect class this
change exists to remove. Documented in CONTRIBUTING.md.
