+++ internal/daemon/sched/cancel_processgroup_5999_test.go
@@ -42,6 +42,21 @@ func fakeGroupAlgoBinary(t *testing.T, body string) {
 	t.Cleanup(func() { groupAlgoChildBinary = prev })
 }
 
+// Budgets for the cancellation contract. These are DEADLOCK budgets, not
+// latency budgets: the failure mode #5999 describes is a runner that blocks
+// FOREVER because a surviving grandchild holds the inherited stdout pipe open,
+// and a grandchild that is never reaped because the kill went to a single pid
+// instead of the process group. Neither failure gets faster on a fast machine,
+// so the budgets are sized to never fire on a busy one — the original 5s/2s
+// pair produced a false red on a loaded macOS runner and passed on retry.
+//
+// Do not tighten these to "assert cancellation is fast". If cancellation
+// latency ever needs a bound, that is a separate perf-tagged test.
+const (
+	cancelUnblockBudget  = 30 * time.Second
+	grandchildReapBudget = 10 * time.Second
+)
+
 // pidIsAlive reports whether pid still exists (signal 0 probe).
 func pidIsAlive(pid int) bool {
 	return syscall.Kill(pid, 0) == nil
@@ -98,12 +113,12 @@ func TestRunSubprocessLinks_CancelKillsGrandchildren(t *testing.T) {
 		if err == nil {
 			t.Fatal("cancelled run returned nil error, want a cancellation error")
 		}
-	case <-time.After(5 * time.Second):
-		t.Fatalf("runner still blocked 5s after cancel — grandchild pid %d alive=%v holding the inherited pipes",
-			pid, pidIsAlive(pid))
+	case <-time.After(cancelUnblockBudget):
+		t.Fatalf("runner still blocked %s after cancel — grandchild pid %d alive=%v holding the inherited pipes",
+			cancelUnblockBudget, pid, pidIsAlive(pid))
 	}
 
-	if !waitPidGone(pid, 2*time.Second) {
+	if !waitPidGone(pid, grandchildReapBudget) {
 		t.Fatalf("grandchild pid %d survived cancellation — the kill did not reach the process group", pid)
 	}
 }
@@ -128,12 +143,12 @@ func TestRunSubprocessGroupAlgo_CancelKillsGrandchildren(t *testing.T) {
 		if err == nil {
 			t.Fatal("cancelled group-algo returned nil error, want a cancellation error")
 		}
-	case <-time.After(5 * time.Second):
-		t.Fatalf("group-algo runner still blocked 5s after cancel — grandchild pid %d alive=%v",
-			pid, pidIsAlive(pid))
+	case <-time.After(cancelUnblockBudget):
+		t.Fatalf("group-algo runner still blocked %s after cancel — grandchild pid %d alive=%v",
+			cancelUnblockBudget, pid, pidIsAlive(pid))
 	}
 
-	if !waitPidGone(pid, 2*time.Second) {
+	if !waitPidGone(pid, grandchildReapBudget) {
 		t.Fatalf("group-algo grandchild pid %d survived cancellation — the kill did not reach the process group", pid)
 	}
 }
